made routes restful
