+++ routes/users.js
@@ -4,6 +4,18 @@ var User = require('../models/user');
 var passport = require('passport');
 var LocalStrategy = require('passport-local').Strategy;
 var connect = require('connect');
+var multer = require('multer');
+var upload = multer({
+    dest: 'public/img'
+});
+
+function loggedIn(req, res, next) {
+    if (req.user) {
+        next();
+    } else {
+        res.redirect('/');
+    }
+}
 
 /* GET users listing. */
 router.get('/', function(req, res, next) {
@@ -45,6 +57,44 @@ router.post('/new', function(req, res, next) {
         req.flash('success_msg', 'You are registered and can now log in');
         res.redirect('/');
     }
+});
+
+router.get('/:_id', loggedIn, function(req, res, next) {
+    User.findById({
+        _id: req.user._id
+    }, function(err, user) {
+        if (err) {
+            return next(err);
+        }
+        res.render('sessions/profile', {
+            user: user
+        });
+    });
+});
+
+router.get('/:_id/edit', loggedIn, function(req, res, next) {
+    res.render('sessions/edit-profile');
+});
+
+router.post('/:_id/edit', upload.single('image'), function(req, res, next) {
+
+    User.findOne({
+        _id: req.user._id
+    }, function(err, user) {
+        if (err) return next(err);
+
+        if (req.body.email) user.email = req.body.email;
+        if (req.body.bio) user.bio = req.body.bio;
+        if (req.body.firstname) user.firstname = req.body.firstname;
+        if (req.body.lastname) user.lastname = req.body.lastname;
+        if (req.file.filename) user.image = req.file.filename;
+
+        user.save(function(err) {
+            if (err) return next(err);
+            req.flash('success', 'Successfully edited your profile');
+            return res.redirect('/users/' + user._id);
+        });
+    });
 });
 
 module.exports = router;
