+++ views/sessions/edit-profile.handlebars
@@ -2,7 +2,7 @@
 
 
 <div>
-  <form method="post" action="/sessions/edit-profile" enctype="multipart/form-data">
+  <form method="post" action="/users/{{user._id}}/edit" enctype="multipart/form-data">
     <input type="file" name="image" id="fileToUpload">
         <br>
     <label for="firstname">
@@ -26,5 +26,3 @@
     <input type="submit" value="Update profile">
   </form>
 </div>
-
-<a href="#"> Home </a>
