+++ views/sessions/profile.handlebars
@@ -24,4 +24,4 @@ Short Bio : {{user.bio}}
 <br>
 <br>
 
-<a href='/sessions/edit-profile/{{user.id}}'>Edit Profile</a>
+<a href='/users/{{user.id}}/edit'>Edit Profile</a>
